iAI/requirements.txt
# iAI - Local AI Chat Assistant
# Python 3.9+ recommended
# For exact versions without conflicts, use: requirements_fixed.txt

flet>=0.28.3
ollama>=0.1.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
peft>=0.4.0
bitsandbytes>=0.40.0
chromadb>=0.3.21
PyPDF2>=3.0.0
pdfminer.six>=20221105
pdfplumber>=0.9.0
python-docx>=0.8.11
Pillow>=9.5.0
requests>=2.31.0
urllib3>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
jsonlines>=3.1.0
python-dotenv>=1.0.0